refactor: alert context and product/order service imports
Updated all admin pages to import useAlert from '@/providers/alert-provider' instead of the old context path. Changed product and order data service imports to use new service files (e.g., product.service, category.service, subcategory.service) for consistency and maintainability. Also updated type definitions for order and product IDs to use string instead of number where appropriate.
